+++ src/app/project/comment-period/comment-period-resolver.service.ts
@@ -27,11 +27,11 @@ export class CommentPeriodResolver implements Resolve<Object> {
     const commentPeriodId = route.paramMap.get('commentPeriodId');
     const projectId = route.parent.params['projId'];
     // force-reload so we always have latest data
-    return forkJoin(
-      from(this.commentPeriodService.getSummaryById(commentPeriodId)),
-      from(this.commentPeriodService.getById(commentPeriodId)),
-      from(this.surveyService.getAllByProjectId(projectId))
-    ).map(([summary, commentPeriod, surveys]) => {
+    return forkJoin([
+      this.commentPeriodService.getSummaryById(commentPeriodId),
+      this.commentPeriodService.getById(commentPeriodId),
+      this.surveyService.getAllByProjectId(projectId)
+    ]).map(([summary, commentPeriod, surveys]) => {
       commentPeriod.summary = summary;
       return { commentPeriod: new CommentPeriod(commentPeriod), surveys: surveys};
     });